Here's the new malloc and free
Again, I've done basic testing already and it seems to work.  The new implementation is actually simpler in some ways, but its nastier to look at cause its got a shit ton of dereferenced pointers and such.  Metadata is stored in two shorts in front of each block of memory; first is the size of the block and second is 1 or 0 for allocated or free respectively.